pyproject.toml
[project]
name = "tessa-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "bcrypt==3.2.2",
    "redis>=6.4.0",
    "alembic>=1.16.4",
    "jinja2>=3.1.6",
    "tenacity>=9.1.2",
    "sib-api-v3-sdk>=7.6.0",
    "pydantic-ai>=0.8.1",
    "python-postmark>=0.6.1",
    "postmarker>=1.0",
]
